Lockdown Browser to mitigate vulnerabilities
Webkit used by the Browser process is highly exposed to security
vulnerabilities. The Browser process should run with as few
permissions as possible to limit the damage of a vulnerability.
The android.permission.WRITE_EXTERNAL_STORAGE permission is used only
for debug and test purposes and should not be included in release
code.
Add the "testOnly" attribute to the uses-permission tag to ensure
that the permission is dropped for release builds (userdebug/user).
This commit has a dependency to the commit in frameworks base which
adds support for the "testOnly" attribute in the uses-permission tag.

Change-Id: I3104fe2135703cbd3ca53f783b59fe15b422053e

Conflicts:

	AndroidManifest.xml

Conflicts:

	src/com/android/browser/BrowserActivity.java

	modified:   AndroidManifest.xml
	modified:   src/com/android/browser/BrowserActivity.java